# specs/001-english-flashcards-web/contracts/api-contract.yaml
openapi: 3.0.3
info:
  title: English Flashcards BFF API
  version: 1.0.0
  description: |
    英语闪卡复习应用的后端接口契约（BFF层），用于导入、学习状态提交、生词流转与看板数据读取。
    已与当前实现路径对齐：/ocr, /cards/enrich, /study/submit, /study/undo, /dashboard/heatmap
servers:
  - url: /api
paths:
  /ocr:
    post:
      summary: 从上传图片提取词条
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
              required: [image]
      responses:
        "200":
          description: 识别成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/RawWordItem"
...
components:
  schemas:
    RawWordItem:
      type: object
      properties:
        word:
          type: string
      required: [word]
...